+++ R/sensor_load.R
@@ -50,7 +50,7 @@ sensor_load <-
     if ( is.null(collection) ) 
       stop("Required parameter 'collection' is missing.")
     
-    dateRange <- .dateRange(startdate, enddate, days, timezone)
+    dateRange <- MazamaCoreUtils::dateRange(startdate, enddate, days, timezone)
     
     # Asssemble monthly archive files ------------------------------------------