+++ local_executables/createPAS_exec.R
@@ -25,10 +25,12 @@ suppressPackageStartupMessages({
 if ( interactive() ) {
   
   # RStudio session
-  opt <- list(outputDir = getwd(),
-              logDir = getwd(),
-              spatialDataDir = "~/Data/Spatial",
-              version = FALSE)  
+  opt <- list(
+    outputDir = getwd(),
+    logDir = getwd(),
+    spatialDataDir = "~/Data/Spatial",
+    version = FALSE
+  )  
   
 } else {
   
@@ -108,13 +110,13 @@ result <- try({
   # Set up MazamaSpatialUtils
   AirSensor::initializeMazamaSpatialUtils(opt$spatialDataDir)
   
-  # Save it with the YYYYmmddHH stamp
+  # Save it with the UTC YYYYmmddHH stamp
   timestamp <- strftime(lubridate::now("UTC"), "%Y%m%d%H", tz="UTC")
   filename <- paste0("pas_", timestamp, ".rda")
   filepath <- file.path(opt$outputDir, filename)
   
   logger.info("Obtaining 'pas' data for %s", timestamp)
-  pas <- pas_loadLatest()
+  pas <- pas_createNew()
   
   logger.info("Writing 'pas' data to %s", filename)
   save(list="pas", file=filepath)